Adding pscsmod and pscsmodorg
Lines 219-222, 234-235, and 248-249 were added to create two new modified themes of particle size in the control section (PSCS). The pscsmod simply updates Psamments in Lithic (shallow) subgroups to 'LITHIC PSAMMENTS' to capture that variation. In pscsmodorg does a few things. It lumps 'PSAMMENTS' in with 'SANDY' since they both are defined exactly the same, and also brings in 'HISTOSOLS' from soil order as 'ORGANIC'. Then to better characterize soil depth we take the LITHIC label from soil subgroup and add that to pscs classes that are not 'CLAYEY' or 'LOAMY', which are shallow already by definition. These changes help better characterize the full landscape of the lower 48 US states, but also better describes soil depth, an important management parameter.